Add SS1 record setFieldText and findLineItemValue.
The `setFieldText` is NOT the same implementation as in NetSuite. There
if you use `setFieldText` and then `getFieldValue` on an optionset field
you will get an integer back, here you will get a string back. This is
fine form my purposes though and is the same as the implementation in
the SS2 module[1].

The `findLineItemValue` is very similar to the other *LineItem* methods
but I thought I would add a test for it anyway for good luck.

[1]: https://github.com/3EN-Cloud/netsumo/blob/0b4ff26c7645f2e5d531c314b06b550d7389def9/modules/SS2/Record.js#L86