SS1: Support arbitrary sublist names.
Adds support for using record sublists other than `item` and
`addressbook`. This shouldn't require any changes in dependant pieces of
code unless they were relying on matching an error message or on
`selectNewLineItem` erroring on unsupported records (it also "upserts"
in the SS2 module).

Most places that would have previously complained about unsupported lines will
now complain about `TypeError: Cannot * property '*' of undefined`. Exceptions
are:
* `selectNewLineItem`: Will create new sublist groups and new currentLine
  groups as appropriate in line with what `selectNewLine` does in the SS2
  module. (I'm not sure what that `addressbookaddress` thing is for but left
  it in for compatibility.)
* `getLineItemCount` it was already returning 0 for unknown sublists so I left
  it that way

I have left `lineItems` and `addressBookLines` as aliases in case they are
being used in existing tests somehow (I don't think that is possible given the
scope though). `lineItems` is still used in `transform()`, I am not sure if
that should be updated and if so what to.